refactor: ProductRequest에서 country 필드 타입을Country(enum) → String으로 변경,매핑 로직을 수정
