+++ src/main/java/com/yeoljeong/tripmate/company/application/service/query/CompanyQueryService.java
@@ -1,7 +1,7 @@
 package com.yeoljeong.tripmate.company.application.service.query;
 
 import com.yeoljeong.tripmate.company.application.dto.result.CompanyResult;
-import com.yeoljeong.tripmate.company.domain.entity.Company;
+import com.yeoljeong.tripmate.company.domain.model.Company;
 import com.yeoljeong.tripmate.company.domain.exception.CompanyErrorCode;
 import com.yeoljeong.tripmate.company.domain.repository.CompanyRepository;
