+++ src/main/java/com/yeoljeong/tripmate/company/infrastructure/persistence/jpa/CompanyJpaRepository.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.company.infrastructure.persistence.jpa;
 
-import com.yeoljeong.tripmate.company.domain.entity.Company;
+import com.yeoljeong.tripmate.company.domain.model.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
