+++ src/main/java/com/yeoljeong/tripmate/product/application/service/query/ProductQueryService.java
@@ -2,7 +2,7 @@ package com.yeoljeong.tripmate.product.application.service.query;
 
 import com.yeoljeong.tripmate.exception.BusinessException;
 import com.yeoljeong.tripmate.product.application.dto.result.ProductResult;
-import com.yeoljeong.tripmate.product.domain.entity.Product;
+import com.yeoljeong.tripmate.product.domain.model.Product;
 import com.yeoljeong.tripmate.product.domain.exception.ProductErrorCode;
 import com.yeoljeong.tripmate.product.domain.repository.ProductRepository;
 import java.util.UUID;
